Move duplicate `LoaderStub` into shared helper folder with new name `FeedLoaderStub`
